# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChatterPay Global API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/ban:
    post:
      operationId: banUser
      tags: [users]
      summary: Ban or unban a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserBody"
      responses:
        "200":
          description: User banned/unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /foreign-clients:
    get:
      operationId: listForeignClients
      tags: [foreign-clients]
      summary: List all foreign clients (admin only)
      responses:
        "200":
          description: List of foreign clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForeignClient"
    post:
      operationId: createForeignClient
      tags: [foreign-clients]
      summary: Create a foreign client account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateForeignClientBody"
      responses:
        "201":
          description: Foreign client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForeignClient"
  /foreign-clients/{clientId}:
    get:
      operationId: getForeignClient
      tags: [foreign-clients]
      summary: Get foreign client by ID
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Foreign client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForeignClient"
...
  /mpesa/deposit:
    post:
      operationId: initiateDeposit
      tags: [mpesa]
      summary: Initiate M-Pesa STK Push deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MpesaDepositBody"
      responses:
        "200":
          description: STK push initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpesaDepositResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mpesa/callback:
    post:
      operationId: mpesaCallback
      tags: [mpesa]
      summary: M-Pesa payment callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Callback processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /notifications/{notifId}/read:
    post:
      operationId: markNotificationRead
      tags: [users]
      summary: Mark a notification as read
      parameters:
        - name: notifId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        kshBalance:
          type: number
        usdtBalance:
          type: number
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        assignedClientsCount:
          type: integer
        hasActiveContract:
          type: boolean
      required: [id, username, kshBalance, usdtBalance, isBanned, createdAt]
    BanUserBody:
      type: object
      properties:
        banned:
          type: boolean
      required: [banned]
    ForeignClient:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        usdtBalance:
          type: number
        country:
          type: string
        isOnline:
          type: boolean
        createdAt:
          type: string
          format: date-time
        assignedUsersCount:
          type: integer
      required: [id, username, usdtBalance, isOnline, createdAt]
    CreateForeignClientBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        usdtBalance:
          type: number
        country:
          type: string
      required: [username, password]
...
    MpesaDepositBody:
      type: object
      properties:
        phone:
          type: string
        amount:
          type: number
          minimum: 1
      required: [phone, amount]
    MpesaDepositResponse:
      type: object
      properties:
        checkoutRequestId:
          type: string
        merchantRequestId:
          type: string
        responseDescription:
          type: string
        message:
          type: string
      required: [message]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [payment_success, contract_activated, client_assigned, withdrawal_approved, withdrawal_rejected]
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, message, isRead, createdAt]